backend/requirements.txt
Django==4.2.*
djangorestframework
django-redis
django-cors-headers
python-dotenv
mysqlclient
drf-spectacular
pytest
pytest-django
pytest-cov
factory-boy
requests
pillow
PyJWT
faker
gunicorn